update releaseDatePrecision
se eliminaron las funciones day(), month(), year() y sus correspondientes chequeos, hacia una nueva clase releaseDatePrecision
